re-fix the intentional misspelling of "MANAGELQ"
(It's to illustrate why the font was chosen. The logo takes this into
account by adding a bit of space before and emboldening the IQ), but
custom tweaks cannot be done for every mention of the name within the
text. Therefore, the font chosen (Noto Sans) helps to prevent people
from reading "MANAGELQ" and read "MANAGEIQ" instead.

Demonstration:
1. ManagelQ
2. ManageIQ

In some typefaces, it is hard to distinguish the two above. Noto Sans is
a (mainly) sans-serif font family that makes a distinction betwen
a lower-case "L" and uppercase "I" so that it removes ambiguation.

And that's why there's a misspelling, showing what we want to avoid.