Add a LWIMIQ post generator
With this revision I mostly wanted a generator that took the pain out of
fiddling with dates and inserting them into GitHub URLs, and also
provide a skeleton that represents the form I typically use every time I
make a LWIMIQ blog post. With this added, an author can type in the
start date of the week concerned and get a blog post skeleton back with
working links.

The intention of this is not to set that form in stone - I love reading
posts that deviate from this formula, or even abandon it completely. But
I personally like having a blog post skeleton that gets me over the
blank canvas stage.