Add links to PRs from main repos, post-split
After the split it became apparent that there's a lot of work that goes
on outside of the "main" repo. So I've added links to what seem to make
up most of the work, and generalized the building of the link so it
should be easier to add/remove (and even read) in the future.