Have lwimiq generator work out the dates
Instead of passing a command line argument to the generator for the
date, have the generator work out what the dates should be for last
week. If you're trying to produce one ahead of time, you can pass the
optional `--current` argument to generate the post for the current
week.