Fix summary blocks in firefox
There is currently a rule in our CSS somewhere that basically does the
following:

  summary { display: block }

This, in Firefox, removes the display of the "arrow" for summary blocks,
though, they still expand just fine.  This change basically makes it so
it undoes that change for the blog specifically and `summary` elements
only.