Remove </br> tags in vignettes post
These get removed by the site engine, and I am too lazy to figure out
how this is done.  There are better ways to add spacing in html
anyway...

Of note, the previous commit that switched the file from `html.md` to
just `.html` had already fixed the issue with `</br>` and `</hr>` tags
showing up in the resulting post, so this was just removing dead weight.