Fix overflow issue with code blocks
The `white-space: pre-wrap` rule for `pre code` blocks was being
inherited from `bootstrap/code`, and was causing word wrap to happen on
long lines.

This looks bad, and is much better to use the already included
`overflow: auto` rule set.

This simply overrides that one rule in `bootstrap/code`, and allows the
rest to function as it has to keep the rest.