Add console color theme
Gives a distinct difference between code highlighted for code snippets,
and representing shell sessions.  By using a dark theme, it gives a
better impression that the viewer is seeing a terminal.

Chose Monokai, somewhat at random, but was mostly looking for something
dark that seemed to compliment what we had already.