Update posts using `<pre><code style=...>` hack
Because of the changes in the previous commit, handling codeblocks with
raw html like this to avoid word wrap is no longer necessary.

I also updated relevant code blocks to use the `console` language
identifier, instead of just using a plain text variant.

Some additional formatting was also done on some select posts to allow
it to work better with the current version of `rouge` that we are using.
Only `$` or `#` is currently accepted as prompts and will highlight
correctly when using that, and words have to be direct snug to those
prompts, which I don't think is as easy to read when compared to `>` as
the prompt char.

In newer versions of `rouge`, we would be able to select this character,
and comments are better recognized, but that should be saved for a more
relevant patch.

Also, some minor indentation was added to the `miqldap-to-sssd.md` for
list elements with code blocks to format better.