[Dockerfile] bundle first; COPY second
First, only copy the contents of the Gemfile/Gemfile.lock to the
container where it will live, and run `bundle install`.

Next, copy the rest of the contents over to the container:

This saves having to do a bundle install after making changes to a site
file over and over again.  This is especially important with the number
of C-extensions that are compiled for this project.