[Dockerfile] Switch to multistage builds
Uses a secondary container for deployment to cut down on final image
size.  Currently using `nginx:alpine`, though this probably can be
changed if a different base distro is preferred.

This cuts down the total image size from 2.24GB to 248MB.

(about half of that is just the certbot-nginx package, and the other
half is the "compiled" site size)

Notable Changes
---------------

- `yum install nginx` is removed (the run container already has this)
- `nginx` configuration is done on the run image
  - Dockerfile actions in the builder container have been moved
- The `COPY` cmd from the builder image uses `--chown` to reduce the
  image size (doing it in a subsequent `RUN` cmd adds causes it to
  basically double the size of the final image)
- The `ENV MIQ_SITE_DEST` needed to be duplicated, though probably could
  have been shared made configurable on the final image.