Add theme switcher
Allows switching between dark/light themes, regardless of OS preference.

In this case, the theme will first be determined by the OS/Browser theme
preference.  Then, when the toggle button is switched, it will override
that value with a css class ('.light-theme' or '.dark-theme') on the top
level `<html>` tag.  This with then switch the theme to the one
preferred by the user, and save that preference in local storage.

When the page is reloaded or a new page is accessed, it will check that
value first before trying to render the page.  This is why some raw JS
code is required in `_head.html` since it is used to render the theme
prior to the full page load.

Font awesome icons for "sun" and "moon" are used for the toggle switch,
but they don't use the built in CSS classes since it is one less thing
that needs to be written into the JS code.