Upgrade Ruby to 3.3 and Jekyll to 4.3
Jekyll comes with the jekyll-sass-converter plugin, so most of the work
in this commit is switching to that gem, which expects an _sass
directory by default. Also, that gem does not support sprockets, so we
need to do some manual handling of the various sass gems to include
their stylesheets and other assets (fonts) in the proper directories.

Fixes #878